Likelihood Factor for Incremental Propensity Score Interventions

Likelihood Factor for Incremental Propensity Score Interventions

Format

R6Class object.

Value

LF_base object.

Constructor

define_lf(LF_ipsi, name, type = "density", likelihood_base, shift_param, treatment_task, control_task, ...)

name

A character, giving the name of the likelihood factor. Should match a node name in the nodes specified by the npsem slot of tmle3_Task.

likelihood_base

A trained Likelihood object, for use in generating a re-scaled likelihood factor.

shift_param

A numeric, specifying the magnitude of the desired incremental propensity score shift (a multiplier of the odds of receiving treatment).

treatment_task

A tmle3_Task object created by setting the intervention to the treatment condition: do(A = 1).

control_task

A tmle3_Task object created by setting the intervention to the control condition: do(A = 0).

...

Not currently used.

Fields

likelihood_base

A trained Likelihood object, for use in generating a re-scaled likelihood factor.

shift_param

A numeric, specifying the magnitude of the desired incremental propensity score shift (a multiplier of the odds of receiving treatment).

treatment_task

A tmle3_Task object created by setting the intervention to the treatment condition: do(A = 1).

control_task

A tmle3_Task object created by setting the intervention to the control condition: do(A = 0).

...

Additional arguments passed to the base class.

References

"Nonparametric Causal Effects Based on Incremental Propensity Score Interventions."

Kennedy, Edward H (2019). Journal of the American Statistical Association. https://doi.org/10.1080/01621459.2017.1422737

"Causal Mediation Analysis for Stochastic Interventions"

Díaz, Iván and Hejazi, Nima S (2020). Journal of the Royal Statistical Society, Series B. https://doi.org/10.1111/rssb.12362