Parameter for the Population Intervention (In)direct Effects

Parameter definition class. See https://doi.org/10.1111/rssb.12362.

Format

R6Class object.

Value

Param_base object.

Constructor

define_param(Param_medshift, shift_param, ..., outcome_node)

observed_likelihood

A Likelihood corresponding to the observed likelihood.

shift_param

A numeric, specifying the magnitude of the desired incremental propensity score shift (a multiplier of the odds of receiving treatment).

...

Not currently used.

outcome_node

A character, giving the name of the node that should be treated as the outcome.

Fields

cf_likelihood

The counterfactual likelihood under the joint stochastic intervention on exposure and mediators.

lf_ipsi

Object derived from LF_base for assessing the joint intervention on exposure and mediators.

treatment_task

A tmle3_Task created by setting the intervention to the treatment condition: do(A = 1).

control_task

A tmle3_Task object created by setting the intervention to the control condition: do(A = 0).

shift_param

A numeric, specifying the magnitude of the desired incremental propensity score shift (a multiplier of the odds of receiving treatment).