Efficient One-Step Estimator

Efficient One-Step Estimator

est_onestep(data, delta, g_learners, e_learners, m_learners, phi_learners,
  w_names, z_names, cv_folds = 10)

Arguments

data
A data.table containing the observed data, with columns in the order specified by the NPSEM (Y, Z, A, W), with column names set appropriately based on the original input data. Such a structure is merely a convenience utility to passing data around to the various core estimation routines and is automatically generated by medshift.

delta
A numeric value indicating the degree of shift in the intervention to be used in defining the causal quantity of interest. In the case of binary interventions, this takes the form of an incremental propensity score shift, acting as a multiplier of the probability with which a given observational unit receives the intervention (EH Kennedy, 2018, JASA; doi:10.1080/01621459.2017.1422737).

g_learners
A Stack (or other learner class that inherits from Lrnr_base), containing a single or set of instantiated learners from sl3, to be used in fitting the propensity score, i.e., g = P(A | W).

e_learners
A Stack (or other learner class that inherits from Lrnr_base), containing a single or set of instantiated learners from sl3, to be used in fitting a propensity score that conditions on the mediators, i.e., e = P(A | Z, W).

m_learners
A Stack (or other learner class that inherits from Lrnr_base), containing a single or set of instantiated learners from sl3, to be used in fitting the outcome regression, i.e., m(A, Z, W).

phi_learners
A Stack (or other learner class that inherits from Lrnr_base), containing a single or set of instantiated learners from sl3, to be used in a regression of a pseudo-outcome on the baseline covariates, i.e., phi(W) = E[m(A = 1, Z, W) - m(A = 0, Z, W) | W).

w_names
A character vector of the names of the columns that correspond to baseline covariates (W). The input for this argument is automatically generated by medshift.

z_names
A character vector of the names of the columns that correspond to mediators (Z). The input for this argument is automatically generated by medshift.

cv_folds
A numeric specifying the number of folds to be created for cross-validation. Use of cross-validation / cross-fitting allows for entropy conditions on the AIPW estimator to be relaxed. Note: for compatibility with make_folds, this value must be greater than or equal to 2; the default is to create 10 folds.