Fit propensity score regression while conditioning on mediators

Fit propensity score regression while conditioning on mediators

fit_e_mech(data, valid_data = NULL, lrnr_stack, z_names, w_names)

Arguments

data
A data.table containing the observed data, with columns in the order specified by the NPSEM (Y, Z, A, W), with column names set appropriately based on the original input data. Such a structure is merely a convenience utility to passing data around to the various core estimation routines and is automatically generated as part of a call to the user-facing wrapper function medshift.

valid_data
A holdout data set, with columns exactly matching those appearing in the preceding argument data, to be used for estimation via cross-fitting. Optional, defaulting to NULL.

lrnr_stack
A Stack object, or other learner class (inheriting from Lrnr_base), containing a single or set of instantiated learners from the sl3 package, to be used in fitting a cleverly parameterized propensity score that includes the mediators, i.e., e = P(A | Z, W).

z_names
A character vector of the names of the columns that correspond to mediators (Z). The input for this argument is automatically generated by a call to the wrapper function medshift.

w_names
A character vector of the names of the columns that correspond to baseline covariates (W). The input for this argument is automatically generated by a call to the wrapper function medshift.