Fit propensity score with incremental stochastic shift intervention

Fit propensity score with incremental stochastic shift intervention

fit_g_mech(data, valid_data = NULL, delta, learners, w_names)

Arguments

data

A data.table containing the observed data, with columns in the order specified by the NPSEM (Y, Z, A, W), with column names set appropriately based on the original input data. Such a structure is merely a convenience utility to passing data around to the various core estimation routines and is automatically generated by medshift.

valid_data

A holdout data set, with columns exactly matching those appearing in the preceding argument data, to be used for estimation via cross-fitting. Optional, defaulting to NULL.

delta

A numeric value indicating the degree of shift in the intervention to be used in defining the causal quantity of interest. In the case of binary interventions, this takes the form of an incremental propensity score shift, acting as a multiplier of the odds with which a given observational unit receives the intervention (EH Kennedy, 2018, JASA; doi:10.1080/01621459.2017.1422737).

learners

A Stack (or other learner class that inherits from Lrnr_base), containing a single or set of instantiated learners from sl3, to be used in fitting the propensity score, i.e., g = P(A | W).

w_names

A character vector of the names of the columns that correspond to baseline covariates (W). The input for this argument is automatically generated by medshift.