One-step or TML estimation of the population intervention direct effect

One-step or TML estimation of the population intervention direct effect

pide(W, A, Z, Y, ids = seq(1, length(Y)), delta, estimator = c("onestep",
  "tmle"), ci_level = 0.95, ...)

Arguments

W
A matrix, data.frame, or similar corresponding to a set of baseline covariates.

A
A numeric vector corresponding to a treatment variable. The parameter of interest is defined as a location shift of this quantity.

Z
A numeric vector, matrix, data.frame, or similar corresponding to a set of mediators (on the causal pathway between the intervention A and the outcome Y).

Y
A numeric vector corresponding to an outcome variable.

ids
A numeric vector of observation-level IDs, allowing for observational units to be related through a hierarchical structure. The default is to assume all units are IID. When repeated IDs are included, both the cross-validation procedures used for estimation and inferential procedures respect these IDs.

delta
A numeric value indicating the degree of shift in the intervention to be used in defining the causal quantity of interest. In the case of binary interventions, this takes the form of an incremental propensity score shift, acting as a multiplier of the odds with which a unit receives the intervention (EH Kennedy, 2018, JASA; doi:10.1080/01621459.2017.1422737).

estimator
The desired estimator of the natural direct effect to be computed. Currently, choices are limited to a substitution estimator, a re-weighted estimator, a one-step estimator, and a targeted minimum loss estimator.

ci_level
A numeric indicating the desired coverage level of the confidence interval to be computed.

...
Additional arguments passed to medshift. Consult the documentation of that function for details.